Add MODULE.bazel file
[bzlmod] is the upcoming package manager for Bazel's external dependencies.

While it's already in the [central registry], it's apparently also helpful to
have such a file in the repository, in case, you need to use overrides.

This change adds such a file, specifically the one from the central repository
with the version bumped.

I can't quite get the tests to work yet; I think there's a bug in how runfiles
trees are created (the rule_python directory contains the version in addition to
the name).

[bzlmod]: https://docs.bazel.build/versions/main/bzlmod.html
[central registry]: https://github.com/bazelbuild/bazel-central-registry