pkg_rpm: Don't have source_date_epoch apply by default; test modularity cleanup (#487)
The logic for `pkg_rpm`'s `source_date_epoch` attribute should apply only when
set.  This change updates the logic so that a negative value for
`source_date_epoch` results in no `--source_date_epoch` being passed to
`make_rpm.py`.  `None` does not seem to be a valid option.

The default is now a negative number, thus disabling this feature unless
explicitly requested.

A regression test was also provided.  While writing it, I noticed a few quirks
in `rpm` that impacted test execution.  These were also addressed, although
perhaps not in the best way.

For more context, see #486.

NOTE: this does not address the similar issue impacting the legacy `pkg_rpm`
rule.  This is discussed in #488.

Fixes #486.