Add 'isexe' check to zip tests.
isexe checks that the file is r?xr?xr?x.  That is, it just
tests for read and execute, without looking at the write bits.
The need is subtle. Some builds are local, and the executables
often come back as 0o755, while others are remote, returning
an immutable object with mode 0o555.