Merge verify_archive_test_lib with verify_archive_test_main.py.tpl (#705)
* Use Label constructor for verify_archive_test macro

One of the `deps` within the `verify_archive_test` macro
depends on a label, which would be incorrectly evaluated
in a caller's BUILD file to be a non-external package.

Add a Label() call to the label to ensure it is evaluated in the
context of rules_pkg. Also add a test to ensure this can be
correctly used from another workspace.

* Merge verify_archive_test_lib into test template

Instead of exporting verify_archive_test_lib as a
public library embed it into the template for the
test itself.