Support for cross-platform RPM package generation
This change introduces proper support for generating RPM packages for the
designated target platform.

The current version of the "pkg_rpm"-rule exposes an "architecture"-option,
which is meant to indicate the package's intended platform.
Unfortunately, this option is being incorrectly utilized. It is being used to
define the "BuildArch"-option in the RPM spec file.
However, as per RPM packaging guide (https://rpm-packaging-guide.github.io/):
  "'BuildArch' should be used if the package is not architecture dependent. For
  example, if written entirely in an interpreted programming language, set this
  to BuildArch: noarch. If not set, the package automatically inherits the
  Architecture of the machine on which it is built".

So, it sounds like it actually has nothing to do with target architecture. It is
supposed to represent either architecture of host machine or "noarch".

The purpose of this patch is to maintain backward compatibility while provide
the solution for the described issue.
It preserves the existing "architecture"-attribute of the "pkg_rpm" rule, as
it might already be in use by someone.
It introduces the "target_architecture"-attribute to the "pkg_rpm" rule, aiming
to utilize it as a "arch"-suffix for the package name and to provide it to the
"rpmbuild"-utility as the value for the "--target" option.