Make pkg_zip compression configurable
Zip offers different compression algorithms and levels. For packaging,
sometimes faster compression speed is more important than size.
With pkg_tar, there is already the option to use a custom compressor.

For pkg_zip these use case are now possible with the exposed zip
compression_level and compression_type arguments in this PR.

There's a new test case that verifies that the compression settings work by
comparing the compressed file sizes.

The default compression level is the same as recently change for pkg_tar in