Allow graceful fallback for compression types in pkg_zip.
If lzma is not available in your python fall back to bzip2.
If bzip2 is not available, fall back to deflate.

A different scenario might be to fail if the desired compression does not exist.
But then we need a way to query for it so we can make the tests work in environments that don't have lzma.