pkg_tar: load attributes for TreeArtifacts
This change teaches pkg_tar to reinvoke file_attributes for each
individual file in a TreeArtifact; this function is used to load the
attributes passed via pkg_tar arguments (like `mode`). Before this
change, there's no way to set individual attributes for files in a
TreeArtifact.

This attempts to maintain the previous behavior of intermediate dirs
being 0o775, unless a custom mode is passed to override it.