Allow substitution of user-defined variables in RPM preamble (#787)
* Allow substitution of user-defined variables in RPM preamble

It's desirable to be able to parameterize some variables in the
preamble such as architecture when RPM packages.  This change enables
variable substitution in the preamble section so that the values may
be injected in this fashion in lieu of only using statically defined
values.

* Deal with mismatched variable definitions

Currently we don't handle things like $(foo or (bar) correctly.
Lacking regex matching, we can compensate for this somewhat by
attempting to find matching pairs of $( and ) and failing if we see
the start of a variable declaration but not its termination.