fix: use appropriate entry type for runfiles that are directories
This fixes a regression at some point from 0.9.1 to 0.10.1, which caused `pkg_tar` to fail when `include_runfiles` was true and some of the sources had data dependencies that were directories.

The code that handles the `include_runfiles` attribute was always assigning ENTRY_IS_FILE as entry type of runfiles. This caused an error in the python code when packaging, as it encountered a directory in an unexpected place.

The fix is just considering ENTRY_IS_TREE types, in the same way as it is done in `_process_pkg_files`.