pkg_install: Support TreeArtifacts. (#885)
* install: Delete unused params from internal functions.

They are so confusing; I would've thought that because they
took the parameters, they would set the user/group for me.
But in fact, _chown_chmod needs to be called afterwards.

* pkg_install: Support TreeArtifacts.

Implementation notes:
For tree artifacts, when creating directories, we mostly follow the
modes set for the whole TreeArtifact, but also +x to allow searching
the directory. This is similar to how pkg_tar etc. handles things.

Link: https://github.com/bazelbuild/rules_pkg/issues/308

* pkg_install: Also add test for modes in TreeArtifacts.

---------

Co-authored-by: HONG Yifan <elsk@google.com>