rpm: fix sub-RPM post_scriptlet and add %postun support (#934)
The existing implementation of `%post` neglected to actually include
the contents of the `%post` scriptlet and simply added a `%post`
section to the spec file without the body, e.g.

    %post sub

    %files sub
    ...

With this change, I've also added support for `%postun` and included
the scriptlet support in the `test_golden_sub_rpm_contents` test case.