Provide default user / group IDs for alma9 rpmbuild docker
This is required for better compatibility with host platforms where
the active uid (from id -u) might conflict with the base image system
user ids.

We chose 1000 as something outside the standard uid range, and
conveniently the default uid on Ubuntu systems, ref:
https://refspecs.linuxfoundation.org/LSB_3.0.0/LSB-Core-generic/LSB-Core-generic/uidrange.html

I've also changed the Dockerfile to pick up the TARGETARCH value so
that we get the right bazelisk executable.